requirements.txt
pandas
requests
alive_progress
statsmodels
rdkit-pypi
sklearn
tensorflow
torch
matplotlib
chemplot
